@@ -10,12 +10,12 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.1",
+    "react": "^18.0.0",
     "react-bootstrap": "^1.2.2",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.0.0",
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "react-scripts": "^5.0.1",
     "recharts": "^1.8.5",
     "source-map-explorer": "^2.4.2"
   },
