@@ -1,10 +1,7 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import Router from "./router";
 
-ReactDOM.render(
-  <>
-    <Router />
-  </>,
-  document.querySelector("#root")
-);
+const container = document.getElementById("root");
+const root = createRoot(container); 
+root.render(<Router />);
