@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import { useEffect } from "react";
 import "../styles/home.css";
 import bg from "../images/bg3.jpg";
 
@@ -13,19 +13,17 @@ export default function Home(props) {
     // eslint-disable-next-line
   }, []);
   return (
-    <>
-      <div
-        style={{
-          height: "100vh",
-          background: `url(${bg}) center/cover no-repeat`,
-        }}
-      >
-        <div style={{ height: "100%", background: "rgba(0,0,0,0.5)" }}>
-          <div style={{ paddingTop: "100px" }}>
-            <h1 id="title">REACT Test Site</h1>
-          </div>
+    <div
+      style={{
+        height: "100vh",
+        background: `url(${bg}) center/cover no-repeat`,
+      }}
+    >
+      <div style={{ height: "100%", background: "rgba(0,0,0,0.5)" }}>
+        <div style={{ paddingTop: "100px" }}>
+          <h1 id="title">React Web App</h1>
         </div>
       </div>
-    </>
+    </div>
   );
 }
